Split separate components out into their own files
