+++ packages/block-library/src/cover/cover-placeholder.js
@@ -1,0 +1,39 @@
+/**
+ * WordPress dependencies
+ */
+import { BlockIcon, MediaPlaceholder } from '@wordpress/block-editor';
+import { __ } from '@wordpress/i18n';
+import { cover as icon } from '@wordpress/icons';
+
+/**
+ * Internal dependencies
+ */
+import { ALLOWED_MEDIA_TYPES } from './shared';
+
+export default function CoverPlaceholder( {
+	disableMediaButtons = false,
+	children,
+	onSelectMedia,
+	onError,
+	style,
+} ) {
+	return (
+		<MediaPlaceholder
+			icon={ <BlockIcon icon={ icon } /> }
+			labels={ {
+				title: __( 'Cover' ),
+				instructions: __(
+					'Drag and drop onto this block, upload, or select existing media from your library.'
+				),
+			} }
+			onSelect={ onSelectMedia }
+			accept="image/*,video/*"
+			allowedTypes={ ALLOWED_MEDIA_TYPES }
+			disableMediaButtons={ disableMediaButtons }
+			onError={ onError }
+			style={ style }
+		>
+			{ children }
+		</MediaPlaceholder>
+	);
+}
